vc_models/requirements.txt
    torch >= 1.10.2
    torchvision >= 0.11.3
    timm==0.6.11
    hydra-core
    wandb>=0.13
    requests